pyproject.toml
[tool.poetry]
name = "poetry-versions-plugin"
version = "0.29.0"
description = ""
authors = ["xy.kong <xy.kong@gmail.com>"]
readme = "README.md"
packages = [{ include = "poetry_versions_plugin" }]

[tool.poetry.dependencies]
python = "^3.11"
gitpython = "^3.1.43"
poetry = "^1.8.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"


[tool.versions]
branch = "master"
commit = "fb01329"
commit_count = 2
is_dirty = true
datetime = "2024-11-17 01:32:48"
version = "0.29.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."poetry.plugin"]
versions = "poetry_versions_plugin.plugin:VersionsPlugin"

[tool.poetry.plugins."poetry.application.plugin"]
versions = "poetry_versions_plugin.plugin:VersionsApplicationPlugin"
